Update errors from validators to not include the location of the error. (#31)
By passing "0" as the second parameter to error, the location of the error is not included.  For validators (and most other errors, IMO), that information is not really needed...the error message itself is the more interesting part.

Also fixed a small issue with parsing out the location in jwt.lua.  Matching ".*" is a greedy match - which means if the error message had a colon, the first part of the message got stripped as well when putting it into the "reason" field.

For those errors that return an object with a "reason" field, there are no changes.