Correct c-struct 'null' checks
In luajit, a null pointer is distinct from nil, and thus "truthy". That
is to say, `if not c_ffi_struct` will never evaluate to true, even if
the struct is null.

However, a null ffi struct does compare equal to nil, so the new
comparison corrects this issue.

This fixes numerous bugs in this library, including multiple cases of
uninitialized data being passed into functions leading to null pointer
dereferences, buffer overflows, and more.